Change submodule term to point to rancher forked repo. Update submodule dependencies
Add alt action for shell and logs to launch in new window
(rancher/rancher#2360)

Change new window name to _blank

Change isConsole to more generic isPopup

Add padding to shell popup

Add some extra height to logs view when in modal

fix styling on log popup

Add new vm logs and vm shell popup
