+++ .gitmodules
@@ -1,6 +1,6 @@
 [submodule "vendor/term.js"]
 	path = vendor/term.js
-	url = https://github.com/chjj/term.js.git
+	url = https://github.com/rancher/term.js.git
 [submodule "vendor/jquery.initialize"]
 	path = vendor/jquery.initialize
 	url = https://github.com/AdamPietrasiak/jquery.initialize.git
