+++ app/components/container-shell/template.hbs
@@ -1,3 +1,8 @@
+{{#if showProtip}}
+  <div class="text-muted r-mt5 r-mb10" style="font-size: 12px;">
+    ProTip: Hold the {{alternateLabel}} key when opening shell access to launch a new window.
+  </div>
+{{/if}}
 <div class="shell-body">
 </div>
 <div class="footer-actions">
