+++ app/components/modal-console/template.hbs
@@ -1,14 +1,1 @@
-<h2><i class="icon icon-terminal"></i> Console: {{instance.displayName}}</h2>
-<div class="console-body clearfix">
-  <div id="noVNC_screen">
-    <canvas class="console-canvas" width="640" height="20">
-      Canvas not supported.
-    </canvas>
-  </div>
-</div>
-
-<div class="footer-actions">
-  <div class="console-status text-muted">{{status}}</div>
-  <button {{action "ctrlAltDelete"}} class="btn btn-default" disabled={{ctrlAltDeleteDisabled}}>Send Ctrl-Alt-Delete</button>
-  <button {{action "cancel"}} class="btn btn-primary">Close</button>
-</div>
+{{vm-console instance=originalModel dismiss=(action "cancel") disconncted=(action "cancel") }}
