+++ app/components/modal-container-logs/template.hbs
@@ -1,21 +1,4 @@
-<h2>
-  <i class="icon icon-file"></i> Logs: {{instance.displayName}}
-</h2>
-{{#if onlyCombinedLog}}
-  <div class="text-muted"><b>Note:</b> Only combined stdout/stderr logs are available for this container because it was run with the TTY (-t) flag.</div>
-{{else}}
-  <ul class="nav nav-tabs" style="border-bottom: 0">
-    <li role="presentation" class="{{if isCombined 'active'}}"><a {{action "changeShow" "combined"}}>Combined</a></li>
-    <li role="presentation" class="{{if isStdOut 'active'}}"><a {{action "changeShow" "stdout"}}>Standard Out</a></li>
-    <li role="presentation" class="{{if isStdErr 'active'}}"><a {{action "changeShow" "stderr"}}>Standard Error</a></li>
-  </ul>
-{{/if}}
-<pre class="log-body {{unless stdErrVisible 'hide-stderr'}} {{unless stdOutVisible 'hide-stdout'}}">
-</pre>
-<div class="footer-actions">
-  <div class="console-status text-muted">{{status}}</div>
-  <button {{action "scrollToTop"}} class="btn btn-default">Scroll to Top</button>
-  <button {{action "scrollToBottom"}} class="btn btn-default">Scroll to Bottom</button>
-  <button {{action "clear"}} class="btn btn-default">Clear Screen</button>
-  <button {{action "cancel"}} class="btn btn-primary">Close</button>
-</div>
+{{container-logs
+instance=originalModel
+dismiss=(action 'cancel')
+}}
