+++ app/components/vm-console/template.hbs
@@ -1,0 +1,19 @@
+<h2><i class="icon icon-terminal"></i> Console: {{instance.displayName}}</h2>
+{{#if showProtip}}
+  <div class="text-muted r-mt5 r-mb10" style="font-size: 12px;">
+    ProTip: Hold the {{alternateLabel}} key when opening shell access to launch a new window.
+  </div>
+{{/if}}
+<div class="console-body clearfix">
+  <div id="noVNC_screen">
+    <canvas class="console-canvas" width="640" height="20">
+      Canvas not supported.
+    </canvas>
+  </div>
+</div>
+
+<div class="footer-actions">
+  <div class="console-status text-muted">{{status}}</div>
+  <button {{action "ctrlAltDelete"}} class="btn btn-default" disabled={{ctrlAltDeleteDisabled}}>Send Ctrl-Alt-Delete</button>
+  <button {{action "cancel"}} class="btn btn-primary">Close</button>
+</div>
