+++ app/console-vm/template.hbs
@@ -1,0 +1,3 @@
+<div class="r-mt20 r-mr20 r-mb20 r-ml20">
+  {{vm-console instance=model showProtip=false dismiss=(action "cancel") disconncted=(action "cancel") }}
+</div>
