+++ app/console/template.hbs
@@ -1,0 +1,13 @@
+<div class="shell-popup">
+  <h2><i class="icon icon-terminal"></i> Shell: {{model.displayName}}</h2>
+  {{#container-shell
+    instance=model
+    showProtip=false
+    cols=80
+    rows=24
+    dismiss=(action "cancel")
+    disconncted=(action "cancel")
+  }}
+    <button {{action "cancel"}} class="btn btn-primary">Close</button>
+  {{/container-shell}}
+</div>
