+++ app/container-log/template.hbs
@@ -1,0 +1,8 @@
+<div class="r-mr20 r-ml20 r-mt20 r-mb20">
+  {{container-logs
+    instance=model
+    showProtip=false
+    logHeight=185
+    dismiss=(action 'cancel')
+  }}
+</div>
