+++ app/styles/pages/_container.scss
@@ -31,3 +31,6 @@
   margin: 10px 0;
 }
 
+.shell-popup {
+  padding: 20px;
+}
