+++ app/vm-log/controller.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+import Console from 'ui/mixins/console';
+
+export default Ember.Controller.extend(Console, {
+});
